uploads rectified

+++ app/routes/admin/view_applications.py
@@ -1,4 +1,4 @@
-from flask import render_template, redirect, request, flash, url_for
+from flask import render_template, redirect, request, flash, url_for, session
 from flask_login import login_required
 from app import app, db, mail
 from app.models.admin import Course
@@ -7,56 +7,68 @@ from flask_mail import Message
 
 
 @app.route('/admin/view_courses')
-@login_required  # Ensure that only logged-in admins can access this page
 def view_courses():
-    # Retrieve a list of courses from the database
-    courses = Course.query.all()
-    return render_template('admin/view_courses.html', courses=courses)
+    if session.get('admin'):
+        # Retrieve a list of courses from the database
+        courses = Course.query.all()
+        return render_template('admin/view_courses.html', courses=courses)
+    else:
+        flash('You are not logged in as an admin. No access to this page', 'danger')
+        return redirect(url_for('admin_login'))
 
 @app.route('/admin/view_applications/<int:course_id>')
-@login_required
 def view_applications(course_id):
-    # Retrieve a list of applications for the selected course
-    course = Course.query.get(course_id)
-    applications = Applicant.query.filter_by(course_id=course_id).all()
-    return render_template('admin/view_applications.html', course=course, applications=applications)
+    if session.get('admin'):
+        # Retrieve a list of applications for the selected course
+        course = Course.query.get(course_id)
+        applications = Applicant.query.filter_by(course_id=course_id).all()
+        return render_template('admin/view_applications.html', course=course, applications=applications)
+    else:
+        flash('You are not logged in as an admin. No access to this page', 'danger')
+        return redirect(url_for('admin_login'))
 
 
 
 @app.route('/admin/view_application/<int:applicant_id>', methods=['GET', 'POST'])
 def admin_view_application(applicant_id):
-    # Fetch the applicant's data from the database based on the applicant_id
-    applicant = Applicant.query.get_or_404(applicant_id)
+    if session.get('admin'):
+        
+        # Fetch the applicant's data from the database based on the applicant_id
+        applicant = Applicant.query.get_or_404(applicant_id)
 
     
 
-    if request.method == 'POST':
+        if request.method == 'POST':
         # Get the admin's action from the form
-        admin_action = request.form.get('admin-action')
+            admin_action = request.form.get('admin-action')
 
-        if applicant.status in ['approved', 'rejected']:
-            flash('The status of this application has already been updated and cannot be changed.', 'danger')
-            return redirect(url_for('admin_dashboard'))
+            if applicant.status in ['approved', 'rejected']:
+                flash('The status of this application has already been updated and cannot be changed.', 'danger')
+                return redirect(url_for('admin_dashboard'))
 
         # Handle the admin's decision, such as updating the status in the database
-        if admin_action == 'accept':
-            applicant.status = 'approved'
-            db.session.commit()
-            send_email_to_applicant(applicant, 'Congratulations! Welcome To Bwave ICT', 'user/application_approved_mail.html')
-        elif admin_action == 'reject':
-            # Delete the application and its associated data from the database
-            db.session.delete(applicant)
-            db.session.commit()
-            send_email_to_applicant(applicant, 'Application Rejected', 'user/application_rejected_mail.html')
-            flash('Application has been rejected and deleted.', 'success')
+            if admin_action == 'accept':
+                applicant.status = 'approved'
+                db.session.commit()
+                send_email_to_applicant(applicant, 'Congratulations! Welcome To Bwave ICT', 'user/  application_approved_mail.html')
+            elif admin_action == 'reject':
+                # Delete the application and its associated data from the database
+                db.session.delete(applicant)
+                db.session.commit()
+                send_email_to_applicant(applicant, 'Application Rejected', 'user/application_rejected_mail.html')
+                flash('Application has been rejected and deleted.', 'success')
+                return redirect(url_for('admin_dashboard'))
+
+            # You can also perform additional actions, such as sending notifications
+
+            flash('Application status updated successfully!', 'success')
             return redirect(url_for('admin_dashboard'))
-
-        # You can also perform additional actions, such as sending notifications
-
-        flash('Application status updated successfully!', 'success')
-        return redirect(url_for('admin_dashboard'))
-
-    return render_template('admin/view_application.html', applicant=applicant)
+        return render_template('admin/view_application.html', applicant=applicant)
+    
+    else:
+        flash('You are not logged in as an admin. No access to this page', 'danger')
+        return redirect(url_for('admin_login'))
+
